LCD: Edge2: Compatible with new ts050 and old ts050 and ts101
